+++ app/Http/Controllers/UserController.php
@@ -9,6 +9,7 @@ use App\User;
 use Illuminate\Http\Request;
 use Illuminate\Support\Str;
 use Illuminate\Validation\Rule;
+use Symfony\Component\HttpFoundation\Response;
 
 class UserController extends Controller
 {
@@ -21,7 +22,12 @@ class UserController extends Controller
      */
     public function index()
     {
-        return UserResource::collection(User::all());
+        abort_if(
+            \Gate::denies('users_' . 'access'),
+            Response::HTTP_FORBIDDEN,
+            '403 Forbidden'
+        );
+        return UserResource::collection(User::where('id', '<>', 1)->get());
     }
 
     /**
@@ -32,15 +38,20 @@ class UserController extends Controller
      */
     public function store(Request $request)
     {
+        abort_if(
+            \Gate::denies('users_' . 'store'),
+            Response::HTTP_FORBIDDEN,
+            '403 Forbidden'
+        );
         $upload_file = null;
 
-        if ($request->file('profile_picture')){
-            $upload_file = $this->validateFile($request,'image_name','profile_picture', $checks = ['image'],true);
+        if ($request->file('profile_picture')) {
+            $upload_file = $this->validateFile($request, 'image_name', 'profile_picture', $checks = ['image'], true);
         }
 
         $request->merge([
-            'role' => json_decode($request->role,true),
-            'permissions' => json_decode($request->permissions,true),
+            'role' => json_decode($request->role, true),
+            'permissions' => json_decode($request->permissions, true),
         ]);
 
         $rules = [
@@ -86,13 +97,23 @@ class UserController extends Controller
         return $upload_file;
     }
 
-    public function show($id)
+    public function show(Request $request, int $id)
     {
+        abort_if(
+            \Gate::denies('users_' . 'show') || ($request->user()->id !== 1 && $id === 1),
+            Response::HTTP_FORBIDDEN,
+            '403 Forbidden'
+        );
         return UserResource::collection(User::where('id', $id)->get());
     }
 
     public function removeProfilePicture(Request $request)
     {
+        abort_if(
+            \Gate::denies('users_' . 'update') || ($request->user_id === 1 && $request->user()->id !== 1),
+            Response::HTTP_FORBIDDEN,
+            '403 Forbidden'
+        );
         $user_model = User::findOrFail($request->user_id);
 
         if ($user_model->file_id === null) {
@@ -114,8 +135,13 @@ class UserController extends Controller
      * @param User $user
      * @return \Illuminate\Http\Response
      */
-    public function update(Request $request, $id)
+    public function update(Request $request, int $id)
     {
+        abort_if(
+            \Gate::denies('users_' . 'update') || ($id === 1 && $request->user()->id !== 1),
+            Response::HTTP_FORBIDDEN,
+            '403 Forbidden'
+        );
         $user = User::findOrFail($id);
         $upload_file = null;
 
@@ -161,7 +187,7 @@ class UserController extends Controller
         if (gettype($upload_file) === 'object') {
             $upload_file = $this->updateProfilePictureInfo($upload_file, $user);
             $file = $this->process_file($user, $upload_file);
-        } elseif ($upload_file === 'array') {
+        } elseif (gettype($upload_file) === 'array') {
             return response()->json($upload_file, 422);
         }
 
@@ -174,8 +200,16 @@ class UserController extends Controller
      * @param  int  $id
      * @return int
      */
-    public function destroy($id)
+    public function destroy(Request $request, int $id)
     {
+        $is_superadmin_id = $id === 1;
+        $is_self_delete = $request->user()->id === $id;
+        $checker = $is_self_delete || $is_superadmin_id;
+        abort_if(
+            \Gate::denies('users_' . 'destroy') || $checker,
+            Response::HTTP_FORBIDDEN,
+            '403 Forbidden'
+        );
         return User::destroy($id);
     }
 }
